New changes to app.js and updates to routes.js
